// types/event.ts
export type Event = {
  slug: string;
  title: string;
  venue: string;
  city: string;
  state: string;
  address: string;
  date: string;
  time: string;
  speaker: string;
  theme: string;
  image: string;
  imageAlt: string;
  description: string;
  contactEmail: string;
  registrationUrl?: string;
};